Updated `nomeFileSave` value and enhanced `txbResposta` update
Updated the value of `nomeFileSave` to a new FTP path. Added code to calculate the position of the last slash in the `nomeFileSave` string using the `LastIndexOf` method. Enhanced the update of the `txbResposta` text box by setting the text to the `nameFile` string, adding a new line, and appending the substring of `nomeFileSave` up to the last slash.